+++ .ambient-package-update/metadata.py
@@ -25,32 +25,7 @@ METADATA = PackageMetadata(
     license_year=2023,
     development_status="5 - Production/Stable",
     has_migrations=False,
-    readme_content=ReadmeContent(
-        tagline="""Welcome to the **django-pony-express** - class-based emails for Django shipping with a full test
-suite.
-
-Similar to class-based view in Django core, this package provides a neat, DRY and testable (!) way to handle your
-emails in Django.""",
-        content="""## Features
-
-* Class-based structure for emails
-   * Avoid duplicate low-level setup
-   * Utilise inheritance and OOP benefits
-   * No duplicated templates for HTML and plain-text
-* Test suite to write proper unit-tests for your emails
-   * Access your test outbox like a Django queryset
-
-## Etymology
-
-> The Pony Express was an American express mail service that used relays of horse-mounted riders. [...] During its
-> 18 months of operation, the Pony Express reduced the time for messages to travel between the east and west US
-> coast to about 10 days.
->
-> https://en.wikipedia.org/wiki/Pony_Express
-
-The name of this package combines the Django mascot (a pony) with a once quite successful mail service in the US.
-Ingenious, right?""",
-    ),
+    readme_content=ReadmeContent(uses_internationalisation=True),
     dependencies=[
         "Django>=3.2",
         "html2text>=2020.1.16",