+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changelog
 
 * *2.1.2* (2024-06-14)
-  * Updates via ambient-package updater
+  * Internal updates via `ambient-package-update`
 
 * *2.1.1* (2024-05-31)
   * Changed log-level to "info" for successful dispatching